recruiterbrain-voice/requirements.txt
# ============================================
# RecruiterBrain Voice Interview System
# Complete Dependencies - Production Ready
# ============================================

# ===== WEB FRAMEWORK =====
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# ===== LIVEKIT (Voice Infrastructure) =====
livekit==0.12.0
livekit-agents==0.8.3
livekit-plugins-deepgram==0.6.0
livekit-plugins-google==0.6.0
livekit-plugins-openai==0.7.1
livekit-plugins-silero==0.6.0

# ===== DATABASE =====
# PostgreSQL
sqlalchemy[asyncio]==2.0.23
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Milvus
pymilvus==2.3.4

# ===== EMBEDDINGS (E5-Base-V2) =====
sentence-transformers==2.2.2
torch>=2.2.0
transformers==4.35.2

# ===== SPEECH SERVICES =====
# Deepgram STT
deepgram-sdk==3.2.7

# Google Cloud TTS (WaveNet/Neural2)
google-cloud-texttospeech==2.16.0
google-auth==2.25.2

# ===== LLM & AI =====
openai>=1.35.0,<2.0.0
anthropic==0.8.1
tiktoken==0.5.2

# ===== TELEPHONY =====
telnyx==2.1.1

# ===== ASYNC & CONCURRENCY =====
aiohttp==3.9.1
aioboto3==12.1.0
aiofiles==23.2.1

# ===== UTILITIES =====
httpx==0.25.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
email-validator==2.1.0

# ===== LOGGING & MONITORING =====
structlog==23.2.0
loguru==0.7.2
prometheus-client==0.19.0
psutil==5.9.6

# ===== TESTING =====
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-timeout==2.2.0
pytest-cov==4.1.0
httpx==0.25.2

# ===== DEVELOPMENT =====
black==23.12.1
flake8==7.0.0
mypy==1.8.0
ipython==8.18.1
